+++ example/poisson_gmg/poisson_package.cpp
@@ -18,21 +18,63 @@
 #include <string>
 #include <vector>
 
+#include <bvals/boundary_conditions_generic.hpp>
 #include <coordinates/coordinates.hpp>
 #include <parthenon/driver.hpp>
 #include <parthenon/package.hpp>
+#include <solvers/bicgstab_solver.hpp>
+#include <solvers/mg_solver.hpp>
 #include <solvers/solver_utils.hpp>
 
 #include "defs.hpp"
 #include "kokkos_abstraction.hpp"
+#include "poisson_equation.hpp"
 #include "poisson_package.hpp"
 
 using namespace parthenon::package::prelude;
 using parthenon::HostArray1D;
 namespace poisson_package {
 
+using namespace parthenon;
+using namespace parthenon::BoundaryFunction;
+// We need to register FixedFace boundary conditions by hand since they can't
+// be chosen in the parameter input file. FixedFace boundary conditions assume
+// Dirichlet booundary conditions on the face of the domain and linearly extrapolate
+// into the ghosts to ensure the linear reconstruction on the block face obeys the
+// chosen boundary condition. Just setting the ghost zones of CC variables to a fixed
+// value results in poor MG convergence because the effective BC at the face
+// changes with MG level.
+
+// Build type that selects only variables within the poisson namespace. Internal solver
+// variables have the namespace of input variables prepended, so they will also be
+// selected by this type.
+struct any_poisson : public parthenon::variable_names::base_t<true> {
+  template <class... Ts>
+  KOKKOS_INLINE_FUNCTION any_poisson(Ts &&...args)
+      : base_t<true>(std::forward<Ts>(args)...) {}
+  static std::string name() { return "poisson[.].*"; }
+};
+
+template <CoordinateDirection DIR, BCSide SIDE>
+auto GetBC() {
+  return [](std::shared_ptr<MeshBlockData<Real>> &rc, bool coarse) -> void {
+    using namespace parthenon;
+    using namespace parthenon::BoundaryFunction;
+    GenericBC<DIR, SIDE, BCType::FixedFace, any_poisson>(rc, coarse, 0.0);
+  };
+}
+
 std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin) {
   auto pkg = std::make_shared<StateDescriptor>("poisson_package");
+
+  // Set boundary conditions for Poisson variables
+  using BF = parthenon::BoundaryFace;
+  pkg->UserBoundaryFunctions[BF::inner_x1].push_back(GetBC<X1DIR, BCSide::Inner>());
+  pkg->UserBoundaryFunctions[BF::inner_x2].push_back(GetBC<X2DIR, BCSide::Inner>());
+  pkg->UserBoundaryFunctions[BF::inner_x3].push_back(GetBC<X3DIR, BCSide::Inner>());
+  pkg->UserBoundaryFunctions[BF::outer_x1].push_back(GetBC<X1DIR, BCSide::Outer>());
+  pkg->UserBoundaryFunctions[BF::outer_x2].push_back(GetBC<X2DIR, BCSide::Outer>());
+  pkg->UserBoundaryFunctions[BF::outer_x3].push_back(GetBC<X3DIR, BCSide::Outer>());
 
   int max_poisson_iterations = pin->GetOrAddInteger("poisson", "max_iterations", 10000);
   pkg->AddParam<>("max_iterations", max_poisson_iterations);
@@ -77,60 +119,60 @@ std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin) {
   Real err_tol = pin->GetOrAddReal("poisson", "error_tolerance", 1.e-8);
   pkg->AddParam<>("error_tolerance", err_tol);
 
+  Real res_tol = pin->GetOrAddReal("poisson", "residual_tolerance", 1.e-8);
+  pkg->AddParam<>("residual_tolerance", res_tol);
+
   bool fail_flag = pin->GetOrAddBoolean("poisson", "fail_without_convergence", false);
   pkg->AddParam<>("fail_without_convergence", fail_flag);
 
   bool warn_flag = pin->GetOrAddBoolean("poisson", "warn_without_convergence", true);
   pkg->AddParam<>("warn_without_convergence", warn_flag);
 
-  // res_err enters a multigrid level as the residual from the previous level, which
-  // is the rhs, and leaves as the solution for that level, which is the error for the
-  // next finer level
-  auto te_type = Metadata::Cell;
-  if (GetTopologicalType(te) == parthenon::TopologicalType::Node) {
-    te_type = Metadata::Node;
-  } else if (GetTopologicalType(te) == parthenon::TopologicalType::Edge) {
-    te_type = Metadata::Edge;
-  } else if (GetTopologicalType(te) == parthenon::TopologicalType::Face) {
-    te_type = Metadata::Face;
-  }
-  using namespace parthenon::refinement_ops;
-  auto mres_err = Metadata({te_type, Metadata::Independent, Metadata::FillGhost,
-                            Metadata::GMGRestrict, Metadata::GMGProlongate});
-  mres_err.RegisterRefinementOps<ProlongateSharedLinear, RestrictAverage>();
-  pkg->AddField(res_err::name(), mres_err);
+  bool use_exact_rhs = pin->GetOrAddBoolean("poisson", "use_exact_rhs", false);
+  pkg->AddParam<>("use_exact_rhs", use_exact_rhs);
 
+  PoissonEquation eq;
+  eq.do_flux_cor = flux_correct;
+
+  parthenon::solvers::MGParams mg_params;
+  mg_params.max_iters = max_poisson_iterations;
+  mg_params.residual_tolerance = res_tol;
+  mg_params.do_FAS = do_FAS;
+  mg_params.smoother = smoother_method;
+  parthenon::solvers::MGSolver<u, rhs, PoissonEquation> mg_solver(pkg.get(), mg_params,
+                                                                  eq);
+  pkg->AddParam<>("MGsolver", mg_solver, parthenon::Params::Mutability::Mutable);
+
+  parthenon::solvers::BiCGSTABParams bicgstab_params;
+  bicgstab_params.max_iters = max_poisson_iterations;
+  bicgstab_params.residual_tolerance = res_tol;
+  bicgstab_params.precondition = precondition;
+  parthenon::solvers::BiCGSTABSolver<u, rhs, PoissonEquation> bicg_solver(
+      pkg.get(), bicgstab_params, eq);
+  pkg->AddParam<>("MGBiCGSTABsolver", bicg_solver,
+                  parthenon::Params::Mutability::Mutable);
+
+  using namespace parthenon::refinement_ops;
   auto mD = Metadata(
       {Metadata::Independent, Metadata::OneCopy, Metadata::Face, Metadata::GMGRestrict});
   mD.RegisterRefinementOps<ProlongateSharedLinear, RestrictAverage>();
+  // Holds the discretized version of D in \nabla \cdot D(\vec{x}) \nabla u = rhs. D = 1
+  // for the standard Poisson equation.
   pkg->AddField(D::name(), mD);
 
-  auto mflux_comm = Metadata({te_type, Metadata::Independent, Metadata::FillGhost,
+  auto mflux_comm = Metadata({Metadata::Cell, Metadata::Independent, Metadata::FillGhost,
                               Metadata::WithFluxes, Metadata::GMGRestrict});
   mflux_comm.RegisterRefinementOps<ProlongateSharedLinear, RestrictAverage>();
+  // u is the solution vector that starts with an initial guess and then gets updated
+  // by the solver
   pkg->AddField(u::name(), mflux_comm);
 
-  auto mflux = Metadata(
-      {te_type, Metadata::Independent, Metadata::FillGhost, Metadata::WithFluxes});
-  mflux.RegisterRefinementOps<ProlongateSharedLinear, RestrictAverage>();
-  pkg->AddField(temp::name(), mflux);
-
-  auto m_no_ghost = Metadata({te_type, Metadata::Derived, Metadata::OneCopy});
-  // BiCGSTAB fields
-  pkg->AddField(rhat0::name(), m_no_ghost);
-  pkg->AddField(v::name(), m_no_ghost);
-  pkg->AddField(h::name(), mflux);
-  pkg->AddField(s::name(), m_no_ghost);
-  pkg->AddField(t::name(), m_no_ghost);
-  pkg->AddField(x::name(), m_no_ghost);
-  pkg->AddField(r::name(), m_no_ghost);
-  pkg->AddField(p::name(), m_no_ghost);
-
-  // Other storage fields
-  pkg->AddField(exact::name(), m_no_ghost);
-  pkg->AddField(u0::name(), m_no_ghost);
+  auto m_no_ghost = Metadata({Metadata::Cell, Metadata::Derived, Metadata::OneCopy});
+  // rhs is the field that contains the desired rhs side
   pkg->AddField(rhs::name(), m_no_ghost);
-  pkg->AddField(rhs_base::name(), m_no_ghost);
+
+  // Auxillary field for storing the exact solution when it is known
+  pkg->AddField(exact::name(), m_no_ghost);
 
   return pkg;
 }